OSD Library START
This patch has no code and will not be submitted as any patchset
It is only to keep both the development branch and the patchset branch
based on a common origin. Also to avoid the first-patch bug in
git-format-patch && a point to rebase the all patchset against.

I will tag this patch as START for easy rebasing

Signed-off-by: Boaz Harrosh <bharrosh@panasas.com>