open-osd: lib: An OSD user-mode library
libosd.so is the same osd_initiator API and source compiled in
user-mode. The commands are issued via bsg.ko kernel driver
and it's sg_io_v4 API.

The patch includes the user-mode Makefile and user-mode emulation
of Kernel API used by the osd_initiator code.

It also adds a thin OSD API for user mode applications that abstaracts
OSD device retrival, managment.

In the next patch osd_ktests are also compiled with use
of this library in user-mode with exact same code.

Thanks to John Chandy <john.chandy@uconn.edu> for fixing bsg
translation on newer systems, and missing "::" on short options.

Signed-off-by: Boaz Harrosh <bharrosh@panasas.com>
Fixes-by: John Chandy <john.chandy@uconn.edu>