open-osd: mkexofs: Multi-device support
This patch changes on-disk format, it is preceded with a patch to exofs
that enables multi-device capabilities:
	[PATCH] exofs: Multi-device mirror support

The patch also changes command-line API so old scripts will no longer
work after this patch, not even with a single device.

It is now possible to specify multiple devices and the default raid
array parameters to use in a single file system.
There are two groups of parameters:
* Global group: which specify global array parameters, and can be
  specified in any order at any place in the command line.
* per-device group: which are processed in order. A --dev=/dev/osdX then
  followed by optional per-device parameters, then a next
  --dev=/dev/osdY and so on. The order of devices is very important.
  See inside the patch for CLI documentation.

The Internal processing is now done in two passes:
* First pass: parses command line options, collects all devices and
  other parameters into a cluster structure, and formats any devices
  specified for format. Devices that need formatting are opened
  formatted then closed.
* 2nd pass: Opens all Devices writes mkfs information on all of them and
  closes.

Signed-off-by: Boaz Harrosh <bharrosh@panasas.com>