open-osd: .gitignore => open-osd.gitignore
For mergability with A Kernel tree, the .gitignore file
is not included in the open-osd git.

However it will be needed by users so it is added to git
as open-osd.gitignore. A user will have to (Once):
	[]$ ln -s open-osd.gitignore .gitignore

And all will work well, because the first file that
is ignored is .gitignore

Signed-off-by: Boaz Harrosh <bharrosh@panasas.com>